feat: move logic to room.rule and room-invitation-link.rule, add room-invitation-link authorizable to hold neccessary data, use rule in room-invitation-link UC

+++ apps/server/src/modules/room/api/room-invitation-link.uc.ts
@@ -1,10 +1,7 @@
 import { AuthorizationService } from '@modules/authorization';
 import { RoleName } from '@modules/role';
 import { RoomMembershipService } from '@modules/room-membership';
-import { SchoolService } from '@modules/school';
-import { User } from '@modules/user/repo';
-import { HttpStatus, Inject, Injectable, NotFoundException } from '@nestjs/common';
-import { FeatureDisabledLoggableException } from '@shared/common/loggable-exception/feature-disabled.loggable-exception';
+import { BadRequestException, HttpStatus, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { EntityId } from '@shared/domain/types';
 import { RoomInvitationLink, RoomInvitationLinkUpdateProps } from '../domain/do/room-invitation-link.do';
 import { RoomInvitationLinkService } from '../domain/service/room-invitation-link.service';
@@ -12,8 +9,9 @@ import { ROOM_PUBLIC_API_CONFIG_TOKEN, RoomPublicApiConfig } from '../room.confi
 import { CreateRoomInvitationLinkBodyParams } from './dto/request/create-room-invitation-link.body.params';
 import { RoomInvitationLinkError } from './dto/response/room-invitation-link.error';
 import { RoomInvitationLinkValidationError } from './type/room-invitation-link-validation-error.enum';
-import { RoomOperation, RoomRule } from '@modules/room-membership/authorization/room.rule';
+import { RoomRule } from '@modules/room-membership/authorization/room.rule';
 import { throwForbiddenIfFalse } from '@shared/common/utils/wrap-with-exception';
+import { RoomInvitationLinkRule } from '@modules/room-membership/authorization/room-invitation-link.rule';
 
 @Injectable()
 export class RoomInvitationLinkUc {
@@ -22,15 +20,15 @@ export class RoomInvitationLinkUc {
 		private readonly roomMembershipService: RoomMembershipService,
 		private readonly roomInvitationLinkService: RoomInvitationLinkService,
 		private readonly authorizationService: AuthorizationService,
-		private readonly schoolService: SchoolService,
-		private readonly roomRule: RoomRule
+		private readonly roomRule: RoomRule,
+		private readonly roomInvitationLinkRule: RoomInvitationLinkRule
 	) {}
 
 	public async createLink(userId: EntityId, props: CreateRoomInvitationLinkBodyParams): Promise<RoomInvitationLink> {
 		const user = await this.authorizationService.getUserWithPermissions(userId);
 		const roomAuthorizable = await this.roomMembershipService.getRoomAuthorizable(props.roomId);
 
-		throwForbiddenIfFalse(this.roomRule.can('addMembers', user, roomAuthorizable));
+		throwForbiddenIfFalse(this.roomRule.can('createRoomInvitationLinks', user, roomAuthorizable));
 
 		const roomInvitationLink = await this.roomInvitationLinkService.createLink({
 			...props,
@@ -42,12 +40,11 @@ export class RoomInvitationLinkUc {
 	}
 
 	public async updateLink(userId: EntityId, props: RoomInvitationLinkUpdateProps): Promise<RoomInvitationLink> {
-		const roomInvitationLink = await this.roomInvitationLinkService.findById(props.id);
-		const { roomId } = roomInvitationLink;
 		const user = await this.authorizationService.getUserWithPermissions(userId);
-		const roomAuthorizable = await this.roomMembershipService.getRoomAuthorizable(roomId);
+		const roomInvitationLink = await this.roomInvitationLinkService.findById(props.id);
+		const roomAuthorizable = await this.roomMembershipService.getRoomAuthorizable(roomInvitationLink.roomId);
 
-		throwForbiddenIfFalse(this.roomRule.can('addMembers', user, roomAuthorizable));
+		throwForbiddenIfFalse(this.roomRule.can('updateRoomInvitationLinks', user, roomAuthorizable));
 
 		roomInvitationLink.title = props.title ?? roomInvitationLink.title;
 		roomInvitationLink.isUsableByExternalPersons =
@@ -65,16 +62,22 @@ export class RoomInvitationLinkUc {
 
 	public async deleteLinks(userId: EntityId, linkIds: EntityId[]): Promise<void> {
 		const roomInvitationLinks = await this.roomInvitationLinkService.findByIds(linkIds);
+
 		if (roomInvitationLinks.length !== linkIds.length) {
 			throw new NotFoundException();
 		}
 
 		const roomIds = roomInvitationLinks.map((link) => link.roomId);
 		const uniqueRoomIds = [...new Set(roomIds)];
-		const roomsAddMembersPermission = await this.canOnRooms(userId, uniqueRoomIds, 'addMembers');
-		const hasAddMembersPermissionOnAllRooms = roomsAddMembersPermission.every(Boolean);
 
-		throwForbiddenIfFalse(hasAddMembersPermissionOnAllRooms);
+		if (uniqueRoomIds.length > 1) {
+			throw new BadRequestException('All links must belong to the same room');
+		}
+
+		const user = await this.authorizationService.getUserWithPermissions(userId);
+		const roomAuthorizable = await this.roomMembershipService.getRoomAuthorizable(uniqueRoomIds[0]);
+
+		throwForbiddenIfFalse(this.roomRule.can('deleteRoomInvitationLinks', user, roomAuthorizable));
 
 		await this.roomInvitationLinkService.deleteLinks(linkIds);
 	}
@@ -83,7 +86,7 @@ export class RoomInvitationLinkUc {
 		const user = await this.authorizationService.getUserWithPermissions(userId);
 		const roomAuthorizable = await this.roomMembershipService.getRoomAuthorizable(roomId);
 
-		throwForbiddenIfFalse(this.roomRule.can('addMembers', user, roomAuthorizable));
+		throwForbiddenIfFalse(this.roomRule.can('listRoomInvitationLinks', user, roomAuthorizable));
 
 		const links = await this.roomInvitationLinkService.findLinkByRoomId(roomId);
 
@@ -95,13 +98,16 @@ export class RoomInvitationLinkUc {
 			this.authorizationService.getUserWithPermissions(userId),
 			this.tryGetLink(linkId),
 		]);
+		const authorizable = await this.roomMembershipService.getRoomInvitationLinkAuthorizable(roomInvitationLink);
 
-		await this.checkValidity(roomInvitationLink, user);
+		this.roomInvitationLinkRule.check('useRoomInvitationLinks', user, authorizable);
+
 		const roleName = await this.ensureUserIsInRoom(roomInvitationLink, userId);
 
 		if (roleName === RoleName.ROOMAPPLICANT) {
 			throw new RoomInvitationLinkError(RoomInvitationLinkValidationError.ROOM_APPLICANT_WAITING, HttpStatus.FORBIDDEN);
 		}
+
 		return roomInvitationLink.roomId;
 	}
 
@@ -134,89 +140,5 @@ export class RoomInvitationLinkUc {
 
 	private async changeRoleTo(roomId: EntityId, userId: EntityId, roleName: RoleName): Promise<void> {
 		await this.roomMembershipService.changeRoleOfRoomMembers(roomId, [userId], roleName);
-	}
-
-	private async checkValidity(roomInvitationLink: RoomInvitationLink | undefined, user: User): Promise<void> {
-		if (!roomInvitationLink) {
-			throw new RoomInvitationLinkError(RoomInvitationLinkValidationError.INVALID_LINK, HttpStatus.NOT_FOUND);
-		}
-
-		this.checkRoleValidity(user, roomInvitationLink);
-
-		this.checkLinkActive(roomInvitationLink);
-
-		await this.checkCreatorSchoolRestriction(user, roomInvitationLink);
-
-		await this.checkStudentFromOtherSchool(user, roomInvitationLink);
-	}
-
-	private checkRoleValidity(user: User, roomInvitationLink: RoomInvitationLink): void {
-		const isTeacher = user.getRoles().some((role) => role.name === RoleName.TEACHER);
-		const isStudent = user.getRoles().some((role) => role.name === RoleName.STUDENT);
-		const isExternalPerson = user.getRoles().some((role) => role.name === RoleName.EXTERNALPERSON);
-
-		if (isTeacher) {
-			return;
-		}
-		if (isStudent && roomInvitationLink.isUsableByStudents) {
-			return;
-		}
-		if (isExternalPerson && roomInvitationLink.isUsableByExternalPersons) {
-			this.checkFeatureLinkInvitationExternalPersonsEnabled();
-			return;
-		}
-
-		throw new RoomInvitationLinkError(
-			RoomInvitationLinkValidationError.NOT_USABLE_FOR_CURRENT_ROLE,
-			HttpStatus.FORBIDDEN
-		);
-	}
-
-	private checkLinkActive(roomInvitationLink: RoomInvitationLink): void {
-		if (roomInvitationLink.activeUntil && roomInvitationLink.activeUntil < new Date()) {
-			throw new RoomInvitationLinkError(RoomInvitationLinkValidationError.EXPIRED, HttpStatus.BAD_REQUEST);
-		}
-	}
-
-	private async checkCreatorSchoolRestriction(user: User, roomInvitationLink: RoomInvitationLink): Promise<void> {
-		const isExternalPerson = user.getRoles().some((role) => role.name === RoleName.EXTERNALPERSON);
-		const skipSchoolRestrictionCheck = isExternalPerson && roomInvitationLink.isUsableByExternalPersons;
-		if (skipSchoolRestrictionCheck) return;
-
-		const creatorSchool = await this.schoolService.getSchoolById(roomInvitationLink.creatorSchoolId);
-		if (roomInvitationLink.restrictedToCreatorSchool && user.school.id !== roomInvitationLink.creatorSchoolId) {
-			throw new RoomInvitationLinkError(
-				RoomInvitationLinkValidationError.RESTRICTED_TO_CREATOR_SCHOOL,
-				HttpStatus.FORBIDDEN,
-				creatorSchool.getInfo().name
-			);
-		}
-	}
-
-	private async checkStudentFromOtherSchool(user: User, roomInvitationLink: RoomInvitationLink): Promise<void> {
-		const isStudent = user.getRoles().some((role) => role.name === RoleName.STUDENT);
-		if (user.school.id !== roomInvitationLink.creatorSchoolId && isStudent) {
-			const creatorSchool = await this.schoolService.getSchoolById(roomInvitationLink.creatorSchoolId);
-			throw new RoomInvitationLinkError(
-				RoomInvitationLinkValidationError.CANT_INVITE_STUDENTS_FROM_OTHER_SCHOOL,
-				HttpStatus.FORBIDDEN,
-				creatorSchool.getInfo().name
-			);
-		}
-	}
-
-	private async canOnRooms(userId: EntityId, roomIds: EntityId[], operation: RoomOperation): Promise<boolean[]> {
-		const user = await this.authorizationService.getUserWithPermissions(userId);
-		const roomAuthorizables = await Promise.all(
-			roomIds.map((roomId) => this.roomMembershipService.getRoomAuthorizable(roomId))
-		);
-
-		return roomAuthorizables.map((roomAuthorizable) => this.roomRule.can(operation, user, roomAuthorizable));
-	}
-
-	private checkFeatureLinkInvitationExternalPersonsEnabled(): void {
-		if (!this.config.featureRoomLinkInvitationExternalPersonsEnabled) {
-			throw new FeatureDisabledLoggableException('FEATURE_ROOM_LINK_INVITATION_EXTERNAL_PERSONS_ENABLED');
-		}
 	}
 }
